Temporarily adjust filter script path for gix
`gix` runs commands from the directory from which it is run to
perform a clone, rather than from within the directory being cloned
in as `git` does.

A better solution will be to use an absolute path, but since the
path might contain weird characters like `'`, let's just make sure
this works first.